healthcare/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
requests
whisper
deep-translator
streamlit